Fix size conversion between different functions.
DNNMark uses a size_t for calculating the workspace size, but
passes it down to functions that use it as an int.  This causes
the workspace size to be inverted to very large negative numbers
when the batch sizes are large.  Currently this only affects
fwd_lrn and bwd_lrn when I make the batch size 256+, but it could
affect any of the others given large enough batch sizes.